Revert small change, idk why it broke meshlet HW raster but it did

+++ crates/bevy_pbr/src/meshlet/visibility_buffer_hardware_raster.wgsl
@@ -54,7 +54,7 @@ fn vertex(@builtin(instance_index) instance_index: u32, @builtin(vertex_index) v
 
 @fragment
 fn fragment(vertex_output: VertexOutput) {
-    let frag_coord_1d = u32(vertex_output.position.y * view.viewport.z + vertex_output.position.x);
+    let frag_coord_1d = u32(vertex_output.position.y) * u32(view.viewport.z) + u32(vertex_output.position.x);
 
 #ifdef MESHLET_VISIBILITY_BUFFER_RASTER_PASS_OUTPUT
     let depth = bitcast<u32>(vertex_output.position.z);
